Protect cloud licenses from being removed

See merge request gitlab-org/gitlab!59576

+++ ee/app/services/licenses/destroy_service.rb
@@ -4,10 +4,13 @@ module Licenses
   class DestroyService < ::Licenses::BaseService
     extend ::Gitlab::Utils::Override
 
+    DestroyCloudLicenseError = Class.new(StandardError)
+
     override :execute
     def execute
       raise ActiveRecord::RecordNotFound unless license
       raise Gitlab::Access::AccessDeniedError unless can?(user, :destroy_licenses)
+      raise DestroyCloudLicenseError.new(_('Cloud licenses can not be removed.')) if license.cloud?
 
       license.destroy
     end
